Create RoundFive.js
With this Codewar we are tasked with creating an algorithm to round the new multiples of five. Looking at this I knew I would need something to round and Math. Ceil helped with that. The issue was getting integers like 4 to equal 5 and 7 to equal 10. At first, I thought of just doing n*5, and boom I got you the answer codewars but the results were not what I expected. After writing out the n and its results I noticed that n is being divided by 5 and results from that is rounded up from ceil and then multiple by Five to get you end results of 4 => 5 and 7 => 10!

Thank you!